hw9/requirements.txt
asgiref==3.2.10
azure-common==1.1.26
azure-core==1.11.0
azure-cosmos==4.2.0
azure-nspkg==3.0.2
azure-storage-blob==12.8.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1
idna==2.10
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.0
Pillow==8.1.1
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3

shortuuid~=1.0.1